requirements.txt
gym[atari,accept-rom-license]==0.21.0
numpy==1.23.4
tqdm==4.64.1
tensorflow==2.11.0
tensorflow-probability==0.19.0
pygame==2.3.0
Pillow==9.5.0
opencv-python==4.7.0.72